requirements.txt
# Core dependencies
python-telegram-bot==20.7
imapclient==3.0.1
mail-parser==3.15.0
beautifulsoup4==4.12.2
pydantic==2.5.3
tenacity==8.2.3
python-dotenv==1.0.0
PyYAML==6.0.1

# Additional dependencies for email parsing
lxml==5.0.0
html2text==2020.1.16

# Async support
asyncio==3.4.3

# Logging enhancements (optional)
colorlog==6.8.0